app/requirements.txt
FastAPI
uvicorn
python-dotenv
google-genai
pydantic